+++ packages/selectors/src/internal/parse.test.ts
@@ -38,3 +38,51 @@ test('parse -> serialize -> parse is idempotent for every selector chain', () =>
     { numRuns: PROPERTY_RUNS },
   );
 });
+
+test('parseSelectorChain parses fallback and boolean terms', () => {
+  const chain = parseSelectorChain('id=auth_continue || role=button label="Continue" visible=true');
+  assert.equal(chain.selectors.length, 2);
+  assert.equal(chain.selectors[0]!.terms[0]!.key, 'id');
+  assert.equal(chain.selectors[1]!.terms[2]!.key, 'visible');
+});
+
+test('parseSelectorChain rejects unknown keys and malformed quotes', () => {
+  assert.throws(() => parseSelectorChain('foo=bar'), /Unknown selector key/i);
+  assert.throws(() => parseSelectorChain('label="unclosed'), /Unclosed quote/i);
+  assert.throws(() => parseSelectorChain(''), /cannot be empty/i);
+});
+
+test('parseSelectorChain handles quoted values ending in escaped backslashes', () => {
+  const chain = parseSelectorChain('label="path\\\\" || id=auth_continue');
+  assert.equal(chain.selectors.length, 2);
+  assert.equal(chain.selectors[0]!.terms[0]!.value, 'path\\');
+});
+
+test('parseSelectorChain decodes escaped selector string values', () => {
+  const chain = parseSelectorChain(
+    [
+      'label="Switch\\nMy Community"',
+      'value="A\\tB\\rC\\bD\\fE\\/F"',
+      'id="item_\\u0031\\uD83D\\uDE00"',
+      "text='It\\'s OK'",
+    ].join(' '),
+  );
+
+  assert.equal(chain.selectors[0]!.terms[0]!.value, 'Switch\nMy Community');
+  assert.equal(chain.selectors[0]!.terms[1]!.value, 'A\tB\rC\bD\fE/F');
+  assert.equal(chain.selectors[0]!.terms[2]!.value, `item_1${String.fromCodePoint(0x1f600)}`);
+  assert.equal(chain.selectors[0]!.terms[3]!.value, "It's OK");
+});
+
+test('parseSelectorChain preserves malformed and unknown selector string escapes', () => {
+  const chain = parseSelectorChain('label="bad\\u12" value="keep\\q"');
+
+  assert.equal(chain.selectors[0]!.terms[0]!.value, 'bad\\u12');
+  assert.equal(chain.selectors[0]!.terms[1]!.value, 'keep\\q');
+});
+
+test('parseSelectorChain preserves literal escaped control sequences when double escaped', () => {
+  const chain = parseSelectorChain('label="foo\\\\nbar"');
+
+  assert.equal(chain.selectors[0]!.terms[0]!.value, 'foo\\nbar');
+});
